Move code where it's used
