+++ commands/active.go
@@ -3,12 +3,9 @@ package commands
 import (
 	"errors"
 	"fmt"
-	"os"
-	"strings"
 
 	"github.com/docker/machine/libmachine/host"
 	"github.com/docker/machine/libmachine/persist"
-	"github.com/docker/machine/libmachine/state"
 )
 
 var (
@@ -49,18 +46,4 @@ func getActiveHost(store persist.Store) (*host.Host, error) {
 	}
 
 	return nil, errors.New("Active host not found")
-}
-
-// IsActive provides a single function for determining if a host is active
-// based on both the url and if the host is stopped.
-func isActive(h *host.Host, currentState state.State, url string) (bool, error) {
-	dockerHost := os.Getenv("DOCKER_HOST")
-
-	// TODO: hard-coding the swarm port is a travesty...
-	deSwarmedHost := strings.Replace(dockerHost, ":3376", ":2376", 1)
-	if dockerHost == url || deSwarmedHost == url {
-		return currentState == state.Running, nil
-	}
-
-	return false, nil
 }
